chore: added dependencies for react hooks
